+++ src/atlas/util/Unique.h
@@ -186,7 +186,7 @@ inline uidx_t UniqueLonLat::operator()( const IndexView<int,1>& elem_nodes ) con
   double centroid[2];
   centroid[LON] = 0.;
   centroid[LAT] = 0.;
-  int npts = elem_nodes.shape(0);
+  size_t npts = elem_nodes.shape(0);
   for( size_t jnode=0; jnode<npts; ++jnode )
   {
     centroid[LON] += lonlat( elem_nodes(jnode), LON );
